+++ pueue/daemon/process_handler.py
@@ -204,7 +204,7 @@ class ProcessHandler():
     def start_process(self, key):
         """Start a specific processes."""
         if key in self.processes and key in self.paused:
-            os.kill(self.processes[key].pid, signal.SIGCONT)
+            os.killpg(os.getpgid(self.processes[key].pid), signal.SIGCONT)
             self.queue[key]['status'] = 'running'
             self.paused.remove(key)
             return True
@@ -218,7 +218,7 @@ class ProcessHandler():
     def pause_process(self, key):
         """Pause a specific processes."""
         if key in self.processes and key not in self.paused:
-            os.kill(self.processes[key].pid, signal.SIGSTOP)
+            os.killpg(os.getpgid(self.processes[key].pid), signal.SIGSTOP)
             self.queue[key]['status'] = 'paused'
             self.paused.append(key)
             return True