+++ CHANGELOG.md
@@ -17,3 +17,4 @@ All notable changes to this project will be documented in this file.
 - Removed `key` parameter from `send` command.
 - Wrong daemon response for `kill` command.
 - `stop` or `kill` sends the signal to all processes spawned by the shell process. This bug affected all command strings which caused the subprocess to spawn a `/bin/sh -c {command}` process.
+- `pause` or `start` sends the signal to all processes spawned by the shell process. This bug affected all command strings which caused the subprocess to spawn a `/bin/sh -c {command}` process.